rust-simple: shell: remove improper lib.hiPrio
The `lib.hiPrio` never did anything in any of the package lists
(including `buildInputs`/`nativeBuildInputs`), and the only reason that
worked before was because of the order in which the packages are listed.

Updated the comment to indicate that ordering is the mechanism for
precedence, and removed the ineffective function call.